Bloomberg Update Check
This code will check the Bloomberg Software Website to see if an update
has been released
for an existing terminal installation.Updates can be released randomly,
sometimes it's helpful
to get ahead of the update, and Bloomberg releases links using a
standard format -- though
only allows auto-update through a throttled system using their
application. People don't like slow.

This code will cycle through all 31 days of all months, even months
without 31 days. A file is
created when the update is found, and the code will not produce a result
again until the next
month. You could program the script to e-mail you of pertinent changes
very easily, or for a
task to download the file to be distributed through the network.